+++ crates/bcr-ebill-persistence/src/tests/mod.rs
@@ -172,6 +172,7 @@ pub mod tests {
                 redeemed_funds_available: false,
                 has_requested_funds: false,
                 last_block_time: test_ts(),
+                is_mature: false,
             },
             state: BillState {
                 mint: BillMintState::None,