+++ crates/bcr-ebill-wasm/main.js
@@ -504,6 +504,7 @@ async function triggerBill(t, blank) {
     const now = new Date();
     const issue_date = now.toISOString().split('T')[0];
     const nMonthsLater = new Date(now); // maturity date today end of day
+    // nMonthsLater.setMonth(nMonthsLater.getMonth() + 1);
     const maturity_date = nMonthsLater.toISOString().split('T')[0];
 
     let file_upload_id = document.getElementById("file_upload_id").value || undefined;